+++ src/Nodes/Switch/SmartIotSwitch.hpp
@@ -4,7 +4,6 @@
 #include <Ticker.h>
 #include <SmartIot.h>
 
-
 class SmartIotSwitch : public SmartIotNode  {
     public:
     SmartIotSwitch(const char* id, const char* name, const char* type = "switch", const SmartIotInternals::NodeInputHandler& nodeInputHandler = [](const String& value) { return false; });
@@ -13,8 +12,8 @@ class SmartIotSwitch : public SmartIotNode  {
     void turnOn()   {_turn(true);}
     void turnOff()  {_turn(false);}
     void toggle();
-    void impulse(uint16_t waveMs);
-    void doubleImpulse(uint16_t waveMs,uint16_t waitMs);
+    void impulse(uint32_t waveMs);
+    void doubleImpulse(uint32_t waveMs,uint32_t waitMs);
     bool SwitchHandler(const String& json);
 
     protected:
@@ -27,7 +26,7 @@ class SmartIotSwitch : public SmartIotNode  {
     uint8_t _pin;
     Ticker _ticker1;
     Ticker _ticker2;
-    uint16_t _doubleWaveMs;
+    uint32_t _doubleWaveMs;
     bool _debounceFlag;
     bool _state;
     void _turn(bool state,bool _debounce = true);
